Well done

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        float fahrenheit=(float)(9*temperatureCelsius/5f+32.0);
+        float fahrenheit=(9*temperatureCelsius/5f+32.0f);
         System.out.println(fahrenheit);
     }
 }
